Disabled the stuff to read zmatricies. The z-matricies weren't being used anywhere - they were just being converted to Cartesian coordinates, and the code todo this transformation within OpenBabel seems to be a bit buggy.
The code to read in zmatricies is still there, but just turned off.

Added a small fix so that we can read in dummy atoms of type x.

Added code to read in the optimised Cartesian coordinates when running
under runtype optimize or saddle.

Added a method to write out a GAMESS-UK input file with some simple
commands in it.

Cleaned up the whitespace which had got a little wild.