Make the split out plugins build everywhere.
This seems to break pytest_* tests, and so may need to be reverted. This works
as expected with the C++ tests as far as I can tell. Initially it did not work
as expected due to the setting of OUTPUT_DIRECTORY. This is set globally, and
should not be set for individual targets.