CMake 2.4 compatibility code...
I am not sure any of the dashboards are actually testing against CMake 2.4.8,
and I spotted quite a few places where matching braces are not present. By
setting this CMake variable to true this is not necessary, and so it should
ensure the loop constructs will at least not cause warnings/failures in old
versions of CMake.