Added environment setting commands to pybel build and install commands.
This fix passes CFLAGS set in CMake to setup.py. setup.py is also passed "-I${EIGEN2_INCLUDE_DIR}", which will otherwise not compile if the Eigen2 headers are not placed in a default location.

This fix uses the export environment setting command, which will not work if the shell does not support this ( sh or bash do, but not csh). A better fix is probably in order