mdlformat.cpp: Support reading and writing CT and tet stereo in 0D MOL files. The "S" option was added to turn off this behaviour. perception.cpp: Correction to CisTransFrom0D for subtle problem exposed by roundtripping CAN->0D SDF->CAN.
All current tests pass, but I should add specific tests for this functionality. Roundtripping 100K molecules from ChEMBL gives 20 failures, which all seem to be due to conjugated cis/trans in a ring (c.f. the well-known failure of SMILES in this area). Maybe something more can be done about this, but that's enough for now.