Moved the C99 support headers to windows-msvc
The C99 support headers do not need any library, and should work with
any version of MSVC, whereas the xdr and cairo headers are tied to a
particular version of some libraries that are paired with binary
prebuild libraries.

It should now be possible to build on any MSVC toolchain if all
dependencies are built correctly.