Make pkg-config check more robust and allow manual override
Add a sensible fallback if pkg-config fails to locate the Open Babel include and library directories, and a helpful error message if that fallback also fails.

Allow Open Babel include and library directories to be manually specified as command line options to setup.py using the build_ext command with -I and -L options.