Fix memory leak in obconversion
pLineEndBuf was not getting deallocated.  In order to do this without causing a crash later, it is necessary to set pInStream's rdbuf back to its original value (assuming pInStream doesn't belong to OBConversion).  Otherwie the stream would have a stale pointer.

An even better fix would be to not change the underlying buffer of a user provided stream and instead put the line endings filter on top.