Fix another bug in fingerprinting FP2, #224
When we reach an atom which is connected back the first atom in ring,
it was closing the ring using the bond order of the bond it used
to reach this last atom, not the bond-order of the bond connecting
the last atom to the first one.

I don't think this bug is path dependent, so any representation of the
same molecule should yield the same fingerprint, they will all be wrong
together.