Fixed a bug with using the --fillUC feature
If pOptions is not NULL but no transforms are found, then
the SpaceGroup object (spacegroup) will contain nothing. This commit adds
a check so that pSG gets set to pUC->GetSpaceGroup() when pOptions is not
NULL but there are no transformations. Because of C++ short-circuits in
operators such as &&, we know that pOptions->find() will never be called
if pOptions is NULL.