Revert "Don't cache OB_MODULE_PATH as it can get stuck on the wrong value even if you change . Don't define it with add_definitions, as it's already set in babelconfig.h and it causes a redefinition warning."
This reverts commit 31cb9a3062267cd88f1a2c6d5eb8c10883723233.

This change breaks Python tests for BINDINGS_ONLY build. It is no longer
possible to provide the module path for system OB by overriding
OB_MODULE_PATH. When installing on Gentoo Linux with Portage, it points
to the staging area (sandbox) of the installation image unless it is
overridden to point to the install destination where the already
installed OB resides.

A proper fix for the redefinition warning would be to add define guards
in babelconfig.h (in follow-up patch).

Signed-off-by: Reinis Danne <rei4dan@gmail.com>