Fix reading of uninitialized data.
Valgrind was reporting Conditional jump or move depends on uninitialised value(s)
I am pretty sure that initializing to zeroes is the right thing to do,
but don't really grok the kekulize code so perhaps this is a symptom of
a larger problem.  It's certainly an improvement over reading
uninitialized values.