Definie OpenBabel3_LIBRARY
Every cmake package I've used exports the location of the library as a
variable.  Perhaps this is no longer considered best practice(?) but it
is certainly common and expected practice.

This addes an OpenBabel3_LIBRARY variable to the cmake config file which
is defined to be $<TARGET_FILE:openbabel>

I could not figure out how to get the location of the library relative
OpenBabel3_INSTALL_PREFIX