cmake: use CMakePackageConfigHelpers for config files
Make use of the CMakePackageConfigHelpers (available already in cmake
3.1) to generate the OpenBabel3ConfigVersion.cmake file completely,
and get rid of the local version.

Mostly important, use CMakePackageConfigHelpers also for generating
the bulk of OpenBabel3Config.cmake: the standard boilerplate from cmake
is inserted, and that takes care of properly locating all the various
paths from the installation directories.

Fixes #2264.