Fix for ammonium xyz parsing.
This is a bug report from Stefano Forli on the openbabel-devel list.
When calling obabel with -p the pH model infers formal charges on an xyz
(since the format doesn't support them) but the code was always
incrementing the number of hydrogens if a charge is added, even if there
were already the appropriate number of explicit hydrogens.  Change the
code to call OBAtomAssignTypicalImplicitHydrogens after updating the
formal charge.  Include a regression test.