Fix for bond order detection for azete
This is a bug reported by Madeleine Walz on the openbabel-devel list.  Given a file format without bond orders, the bond orders for azete were not being correctly inferred.  Madeleine reported the bug with Gaussian input, but has been strangely resistent to providing actual example files so the test is with xyz.

The issue is PerceiveBondOrders uses bond angles to estimate
hybridization states and azete has a 90 degree angle (so the N
hybridization is set to zero and double bonds aren't made where they
should).  The fix is a pretty narrow check that probably doesn't generalize to any other molecules.

Regression test included.