Fix bond perception of ring nitrogens.
There are two rules in bondtyp.txt that I don't understand the
justification of and I don't see how they can be correct when you take
into account that they are applied before aromaticity is perceived so
I've removed them. I have included a test case from a PDB ligand that fails one of these
rules.

If there is a good reason to leave these in then we may need to call
AssignFunctionalGroupBonds after pass 5 of PerceiveBondOrders so
aromaticity is set or figure some otherway to narrow the rules to not
apply in inappropriate cases.