Implemented code to support the Alexandria force field.
The Alexandria Chemistry Toolkit for building force fields
will be released shortly. It relies on the OpenBabel software to
support it for generating atom types. In this patch the
forcefieldalexandria is introduced as a child class to forcefieldgaff
in an effort to change as little code as possible.

There are many changes to bondtyp.txt which admittably will be hard to
check. These will affect any force field typing in OpenBabel.