Update pdbqtformat.cpp to match Meeko PDBQT VDW radius
The current tool recommended by the AutoDock team for ligand preparation is Meeko. Currently, OpenBabel is able to produce pdbqt files extremely similar to Meeko, with the key difference being a vdw radius of 0.00 instead of 1.00. I will attach some text of the same molecule prepared with openbabel and with meeko for reference.